feat: singleton deployment improved and latest factory deployments added to constants
