+++ package.json
@@ -9,6 +9,7 @@
   "scripts": {
     "build": "hardhat compile",
     "build:factory": "tsc -p tsconfig.json",
+    "deploy": "hardhat deploy --network",
     "test:js": "hardhat test ./src/*/__tests__/*.spec.ts",
     "test": "hardhat test && yarn test:js",
     "coverage": "hardhat coverage",
