+++ src/factory/constants.ts
@@ -6,7 +6,7 @@ export const CONTRACT_ADDRESSES: Record<number, ContractAddresses> = {
     realityERC20: "",
     bridge: "",
     delay: "",
-    factory: "0x00000b28D83018aD4F2F4e3a0B4ce5271F4EEa77",
+    factory: "0x00000000062c52e29e8029dc2413172f6d619d85",
     exit: "",
     scopeGuard: "",
     circulatingSupply: "",
@@ -16,7 +16,7 @@ export const CONTRACT_ADDRESSES: Record<number, ContractAddresses> = {
     realityERC20: "0xF669cfdf499e48e0B189206eED98C424F5f83dA7",
     bridge: "0x8a53d2762f90711291c7168fDfC776fB3CFFf7ab",
     delay: "0x753dAEDD9b4464D5AC0b46Ec5c54163678b5ecA3",
-    factory: "0x00000b28D83018aD4F2F4e3a0B4ce5271F4EEa77",
+    factory: "0x00000000062c52e29e8029dc2413172f6d619d85",
     exit: "0x793A69fF46ce0f8C825f14e339cd0Dc8De6A8400",
     scopeGuard: "0xfDc921764b88A889F9BFa5Ba874f77607a63b832",
     circulatingSupply: "0xd7a85e7D0813F8440602E243Acb67df3CCeb5a60",
@@ -26,7 +26,7 @@ export const CONTRACT_ADDRESSES: Record<number, ContractAddresses> = {
     realityERC20: "",
     bridge: "",
     delay: "",
-    factory: "0x00000b28D83018aD4F2F4e3a0B4ce5271F4EEa77",
+    factory: "0x00000000062c52e29e8029dc2413172f6d619d85",
     exit: "",
     scopeGuard: "",
     circulatingSupply: "",
@@ -36,7 +36,7 @@ export const CONTRACT_ADDRESSES: Record<number, ContractAddresses> = {
     realityERC20: "",
     bridge: "",
     delay: "",
-    factory: "0x00000b28D83018aD4F2F4e3a0B4ce5271F4EEa77",
+    factory: "0x00000000062c52e29e8029dc2413172f6d619d85",
     exit: "",
     scopeGuard: "",
     circulatingSupply: "",
@@ -47,7 +47,7 @@ export const CONTRACT_ADDRESSES: Record<number, ContractAddresses> = {
     realityERC20: "0xF669cfdf499e48e0B189206eED98C424F5f83dA7",
     bridge: "0x8a53d2762f90711291c7168fDfC776fB3CFFf7ab",
     delay: "0x753dAEDD9b4464D5AC0b46Ec5c54163678b5ecA3",
-    factory: "0x00000b28D83018aD4F2F4e3a0B4ce5271F4EEa77",
+    factory: "0x00000000062c52e29e8029dc2413172f6d619d85",
     exit: "0x793A69fF46ce0f8C825f14e339cd0Dc8De6A8400",
     scopeGuard: "0xfDc921764b88A889F9BFa5Ba874f77607a63b832",
     circulatingSupply: "0xd7a85e7D0813F8440602E243Acb67df3CCeb5a60",
